+++ src/modules/errors.ts
@@ -86,6 +86,10 @@ export const ERRORS = {
     ALPHA_GRADIENT_LOCATION_MISMATCH: {
         type: 'alpha-gradient-location',
         message: 'Different locations for rgb and alpha codes are not currently supported, resulting in altered opacity. For optimal results, alpha should be located directly with rgb.'
+    },
+    UNSUPPORTED_IMAGE_FEATURES: {
+        id: 'LOTTIE_IMAGE_FEATURES_NOT_SUPPORTED',
+        message: '"Mask" or "Matte" is not supported in Image layer'
     }
 };
 
@@ -146,11 +150,6 @@ export const LAYER_ITEM_PROP_ERROR_MAP = {
         id: 'LOTTIE_ERRORS_MATTES',
         type: 'has-mattes',
         message: '"Mattes" property is not fully supported. We only support Matte Mode "Normal" ("tt" = 0) and "Alpha" ("tt" = 1).'
-    },
-    cl: {
-        id: 'LOTTIE_ERRORS_CLASS',
-        type: 'property-class',
-        message: '"Class" property is not fully supported.'
     },
     hix: {
         id: 'LOTTIE_ERRORS_HIX',